requirements.txt
requests
psycopg2-binary
pandas
python-dotenv
datasketch
transformers
torch
streamlit
sqlalchemy
boto3